_reference/pyproject.toml
[project]
name = "quran-srs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmigrate==0.3.0",
    "humanize>=4.12.1",
    "monsterui==1.0.19",
    "numpy==2.2.3",
    "pandas==2.2.3",
    "pytest>=8.4.2",
    "python-fasthtml==0.12.1",
]
[tool.pytest.ini_options]
pythonpath = "."

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-playwright>=0.7.1",
]
